Frequency interpolation functions moved to spectral_interpolation.py
