+++ rabies/analysis_pkg/diagnosis_pkg/diagnosis_functions.py
@@ -214,7 +214,7 @@ def carpetplot(timeseries, ax, frame_mask=None):
 
 def plot_freqs(ax,timeseries, TR, frame_mask):
     if (frame_mask==0).sum()>0: # only simulate data points if some censoring was applied
-        from rabies.confound_correction_pkg.utils import lombscargle_fill
+        from rabies.confound_correction_pkg.spectral_interpolation import lombscargle_fill
         timeseries_ = lombscargle_fill(x=timeseries,time_step=TR,time_mask=frame_mask)
     else:
         timeseries_ = timeseries
