+++ rabies/confound_correction_pkg/utils.py
@@ -500,89 +500,6 @@ def remove_trend(timeseries, frame_mask=None, order=1 , time_interval='0-end', r
     fitted_intercept = W[-1,:] # predicted intercept
 
     return residuals, fitted_intercept
-    
-
-def lombscargle_mathias(t, x, w):
-
-    '''
-    Implementation of Lomb-Scargle periodogram as described in Mathias et al. 2004, 
-    and applied in Power et al. 2014
-    
-    Mathias, A., Grond, F., Guardans, R., Seese, D., Canela, M., & Diebner, H. H. (2004). 
-    Algorithms for spectral analysis of irregularly sampled time series. Journal of Statistical Software, 11(1), 1-27.
-    
-    Power, J. D., Mitra, A., Laumann, T. O., Snyder, A. Z., Schlaggar, B. L., & Petersen, S. E. (2014). 
-    Methods to detect, characterize, and remove motion artifact in resting state fMRI. Neuroimage, 84, 320-341.
-    
-    '''
-
-    if (w == 0).sum()>0:
-        raise ZeroDivisionError()
-
-    # Check input sizes
-    if t.shape[0] != x.shape[0]:
-        raise ValueError("Input arrays do not have the same size.")
-
-    # Create empty array for output periodogram
-    cw = np.empty((len(w)))
-    sw = np.empty((len(w)))
-    theta = np.empty((len(w)))
-
-    w_t = w[:,np.newaxis].dot(t[np.newaxis,:])
-    theta = (1 / (2 * w)) * np.arctan2(
-        np.sin(2 * w_t).sum(axis=1),
-        np.cos(2 * w_t).sum(axis=1))
-
-        
-    wt = (t[:,np.newaxis]-theta[np.newaxis,:])*w
-    c=np.cos(wt)
-    s=np.sin(wt)
-    
-    cw = x.T.dot(c)/(c**2).sum(axis=0)
-    sw = x.T.dot(s)/(s**2).sum(axis=0)
-    
-    return cw,sw,theta
-
-
-def lombscargle_mathias_simulate(t, w, cw, sw,theta):
-    # recover simulated timeseries for a given time vector t
-
-    wt = (t[:,np.newaxis]-theta[np.newaxis,:])*w
-    c=np.cos(wt)
-    s=np.sin(wt)
-    
-    y = c.dot(cw.T)+s.dot(sw.T)
-    return y
-
-
-def lombscargle_fill(x,time_step,time_mask):
-    num_timepoints=len(time_mask)
-    time=np.linspace(time_step,num_timepoints*time_step,num_timepoints)
-
-    low_freq = 0.005
-    high_freq = 1
-    freqs=np.linspace(low_freq,high_freq,1000)
-    w = freqs*2*np.pi
-
-    t = time[time_mask]
-    cw,sw,theta = lombscargle_mathias(t, x, w)
-    
-    # simulate over entire time
-    t=time
-    y = lombscargle_mathias_simulate(t, w, cw, sw,theta)
-    
-    # standardize according to masked data poaxis=0ints    
-    y -= y[time_mask].mean(axis=0)
-    y /= y[time_mask].std(axis=0)
-    
-    # re-scale according to original mean/std
-    y *= x.std(axis=0)
-    y += x.mean(axis=0)
-    
-    y_fill = np.zeros(y.shape)
-    y_fill[time_mask,:] = x
-    y_fill[time_mask==0,:] = y[(time_mask==0)]
-    return y_fill
 
 
 def butterworth(signals, TR, high_pass, low_pass):
@@ -649,7 +566,7 @@ def phase_randomized_regressors(regressors_array, frame_mask, TR):
     of original regressors to make them independent from the true noise."
 
     """
-    from rabies.confound_correction_pkg.utils import lombscargle_fill
+    from rabies.confound_correction_pkg.spectral_interpolation import lombscargle_fill
     num_conf = regressors_array.shape[1]
     randomized_regressors_array = np.zeros(regressors_array.shape)
     for n in range(num_conf):
